Backend/research_pipeline/requirements.txt
# Research Pipeline Dependencies
biopython>=1.81
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
